Clean up various warnings when building under -Wall
Add some gimmicks to prevent fread() and fwrite() warnings about ignoring
the return value. Yes, this is probably not a good thing to do, but doing
something with a non-success return value is going to involve crashing out
or something anyway.

Also fix several warnings about variables used while possibly uninitialized.
The code flow shows that this cannot be the case but initializing them to a
plausible value at declaration time costs pretty much nothing and it makes
gcc happy.

Also caught a use of | instead of || which probably would have caused a
certain check for duplicate declarations in __os9 sections to behave oddly.