Add notes about coding style and submission guidelines.
As became clear from recent conversation, matching submitted code to the
format of the existing code is not an obvious thing to do. (It should be but
apparently it just doesn't occur to people.) Some description of the coding
style is now present in the README.

Additionally, submission guidelines have also been added to make it clear
that patches (diffs) are the preferred means of providing code changes.