Silence tautological comparison warnings
Because "char" may be signed or unsigned depending on the system, comparing
for byte values 128...255 has to be forced as an unsigned comparison to be
portable and clear. Modify two such comparisons in the symbol handling code
to treat the symbol offset as a pointer to unsigned. This quiets
tautological comparison warnings on systems with an 8 bit signed type as the
default char type.