Add map file output and option to suppress local symbols in listings
Add --map option to generate a symbol map in a regular format that can be
used as an import for other tools.

Also add --symbols-nolocals to suppress output of local symbols in the
symbol listing. In other words, only global symbols are listed.

Thanks to Erik G <erik@6809.org> for the patch.