Update gcc6809 patch to work with pragma=newsource
An updated version of the gcc6809 patch is provided. This update removes
some incompatible syntax left over from the original gcc6809 patch that used
as6809. It was doing something that doesn't work with lwtools anyway.
Namely, trying to use the assembler to set the layout used by the linker.
Now thats all back to front when you think about it.

With this update, it is now possible to use the "as" wrapper with
--pragma=newsource enabled and be able to pass assembly code through the C
preprocessor without worrying about extraneous spaces causing problems.