Change to engage forwardrefmax pragma by default
There are enough cases where the forward reference optimization scheme in
lwasm makes assembly painfully slow that it warrants enabling this pragma by
default to avoid bug reports about lwasm being too slow. Which is is when
you have a lot of forward refences or other instruction sizes that don't
reduce on pass one. This gets worse exponentially as the source file size
gets bigger.

The old behaviour is available to anyone who needs or wants it, though. It
does work quite well on smallish files or files that do not have a lot of
abiguous instruction sizes.

It should be noted that this change does not change the correctness of the
code output by lwasm. What it might do is result in slightly larger code if
the assembler is forced to select a 16 bit mode when an 8 bit mode will do.
However, in most cases I've seen, it makes very little difference overall.