Add a fix and an improvement to gcc6809 patch
Thanks to Mark R V Murray <mark@grondar.org> for a fix to gcc6809 which
fixes a lingering as6809ism (= instead of equ and whitespace before the
symbol).

Also change the bit that is supposed to fill space with NULs to use zmb
instead of a sequence of byte and word definitions.