Update "clrq" convenience instruction to be slightly faster
From Doug Masten <doug@dougmasten.com> who says:

This is not a bug fix but a very tiny optimization for the "clrq"
convenience instruction in LWTOOLS. This patch changes it from "clra; clrb;
tfr d,w" to "clrd; clrw". The new instructions now take 2 clock cycles less
for a total of 4 clock cycles.